remove demux post processing code
