+++ cg/apps/demultiplex/sample_sheet/api.py
@@ -7,6 +7,10 @@ from cg.apps.demultiplex.sample_sheet.read_sample_sheet import get_flow_cell_sam
 from cg.apps.demultiplex.sample_sheet.sample_models import FlowCellSample
 from cg.apps.demultiplex.sample_sheet.sample_sheet_creator import SampleSheetCreator
 from cg.apps.demultiplex.sample_sheet.sample_sheet_validator import SampleSheetValidator
+from cg.apps.demultiplex.sample_sheet.utils import (
+    delete_sample_sheet_from_housekeeper,
+    add_and_include_sample_sheet_path_to_housekeeper,
+)
 from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.apps.lims import LimsAPI
 from cg.apps.lims.sample_sheet import get_flow_cell_samples
@@ -22,10 +26,6 @@ from cg.exc import (
     SampleSheetFormatError,
 )
 from cg.io.controller import ReadFile, WriteFile, WriteStream
-from cg.meta.demultiplex.housekeeper_storage_functions import (
-    add_and_include_sample_sheet_path_to_housekeeper,
-    delete_sample_sheet_from_housekeeper,
-)
 from cg.models.run_devices.illumina_run_directory_data import IlluminaRunDirectoryData
 from cg.utils.files import get_directories_in_path, link_or_overwrite_file
